mcp-servers/http_gateway/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "http-gateway-mcp"
version = "0.1.0"
description = "An MCP server that turns any HTTP API into an MCP tool"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.0",
    "httpx>=0.27",
    "tenacity>=8.0",
    "jsonschema>=4.0",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "bcrypt>=4.0",
]

[project.scripts]
http-gateway = "http_gateway.server:main"

[tool.hatch.build.targets.wheel]
packages = ["http_gateway"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "S", "B", "A", "C4", "PT"]
ignore = ["S101"]  # allow assert in tests

[tool.coverage.run]
source = ["http_gateway"]
omit = ["tests/*"]

[tool.coverage.report]
fail_under = 80

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "types-jsonschema>=4.0",
    "types-PyYAML>=6.0",
    "pytest-cov>=5.0",
    "respx>=0.21",
    "ruff>=0.4",
]
